Unfix warning

+++ code/cgame/cg_draw.c
@@ -2297,7 +2297,7 @@ CG_DrawProxWarning
 static void CG_DrawProxWarning( void ) {
 	char s [32];
 	int			w;
-  static float proxTime;
+  static int proxTime;
   static int proxCounter;
   static int proxTick;
